extract TimedStorage to utils
