+++ hivemind/utils/__init__.py
@@ -5,4 +5,5 @@ from hivemind.utils.serializer import *
 from hivemind.utils.mpfuture import *
 from hivemind.utils.threading import *
 from hivemind.utils.grpc import *
+from hivemind.utils.timed_storage import *
 from hivemind.utils.logging import get_logger
